+++ sc-memory/sc-memory/sc_template.hpp
@@ -48,29 +48,46 @@ struct ScTemplateItem
     SetReplacement(name.c_str());
   }
 
-  inline bool IsAddr() const
+  [[nodiscard]] inline bool IsAddr() const
   {
     return (m_itemType == Type::Addr);
   }
 
-  inline bool IsReplacement() const
+  [[nodiscard]] inline bool IsReplacement() const
   {
     return (m_itemType == Type::Replace);
   }
 
-  inline bool IsType() const
+  [[nodiscard]] inline bool IsType() const
   {
     return (m_itemType == Type::Type);
   }
 
-  inline bool IsFixed() const
+  [[nodiscard]] inline bool IsFixed() const
   {
     return IsAddr() || (IsReplacement() && m_addrValue.IsValid());
   }
 
-  inline bool IsAssign() const
+  [[nodiscard]] inline bool IsAssign() const
   {
     return m_itemType == Type::Type;
+  }
+
+  [[nodiscard]] inline bool HasName() const
+  {
+    return !m_name.empty();
+  }
+
+  [[nodiscard]] inline std::string GetPrettyName() const
+  {
+    std::string result;
+    if (HasName())
+    {
+      result.reserve(2 + m_name.size());
+      result = "`" + m_name + "`";
+    }
+
+    return result;
   }
 
   void SetAddr(ScAddr const & addr, sc_char const * replName = nullptr)
@@ -98,15 +115,12 @@ struct ScTemplateItem
       m_name = name;
   }
 
-  Type m_itemType;
+  Type m_itemType = Type::Type;
 
   ScAddr m_addrValue;
   ScType m_typeValue;
   std::string m_name;
 };
-
-//! backward compatibility
-using ScTemplateItemValue = ScTemplateItem;
 
 class ScTemplateTriple
 {
@@ -129,7 +143,7 @@ public:
     m_values[2] = param3;
   }
 
-  ScTemplateTripleItems const & GetValues() const
+  [[nodiscard]] ScTemplateTripleItems const & GetValues() const
   {
     return m_values;
   }
@@ -233,14 +247,14 @@ public:
     return false;
   }
 
-  SC_DEPRECATED(0.9.0, "Don't use this method, it is dangerous. It will be removed in 0.10.0.")
+  [[nodiscard]] SC_DEPRECATED(0.9.0, "Don't use this method, it is dangerous. It will be removed in 0.10.0.")
 
-  _SC_EXTERN ScTemplateItemsToParams GetAll() const
+      _SC_EXTERN ScTemplateItemsToParams GetAll() const
   {
     return m_templateItemsToParams;
   }
 
-  _SC_EXTERN bool IsEmpty() const noexcept
+  [[nodiscard]] _SC_EXTERN bool IsEmpty() const noexcept
   {
     return m_templateItemsToParams.empty();
   }
@@ -299,7 +313,7 @@ public:
       return m_result;
     }
 
-    std::string const & Msg() const
+    [[nodiscard]] std::string const & Msg() const
     {
       return m_msg;
     }
@@ -710,7 +724,7 @@ public:
   }
 
 protected:
-  sc_memory_context const * m_context;
+  sc_memory_context const * m_context = nullptr;
 
   using SearchResults = std::vector<ScAddrVector>;
   SearchResults m_replacementConstructions;