+++ sc-memory/sc-memory/sc_template_search.cpp
@@ -707,7 +707,11 @@ private:
     ScIterator3Ptr it =
         CreateIterator(templateTriple, result.m_replacementConstructions[replacementConstructionIdx], result);
     if (!it || !it->IsValid())
-      SC_THROW_EXCEPTION(utils::ExceptionInvalidState, "During search procedure has been chosen var triple");
+      SC_THROW_EXCEPTION(
+          utils::ExceptionInvalidState,
+          "Fully variable triple was selected during searching by specified sc-template. It is possible that you have "
+          "incorrect sc-template or you can't find constructions in knowledge base using this sc-template. Check "
+          "sc-template.");
 
     size_t checkedCurrentResultEqualTemplateTriplesCount = 0;